Spiking sort of as a proof of concept.
Right now it'll take a hardcoded set of paths to assemblies and generate
signed versions of them. When it encounters an assembly that is signed
in one it's signing it'll change the reference to include the
publickeytoken to maintain the proper references among them. This is a
necessary step because all signed assemblies must have their own
references signed.

Also one thing that's important is to get all the assemblies that are
referencing those assemblies that are signed, otherwise you'll end up
with references to the old-unsigned one and it'll break. Gotta pay
attention to the public key token on your references and error messages.